Add specific controller for status routes
