+++ library/src/main/java/eu/miaplatform/customplugin/springboot/CPConstants.java
@@ -1,0 +1,5 @@
+package eu.miaplatform.customplugin.springboot;
+
+public class CPConstants {
+    public static final String CP_REQUEST = "CP_REQUEST";
+}
