+++ library/src/main/java/eu/miaplatform/customplugin/springboot/CPController.java
@@ -2,7 +2,6 @@ package eu.miaplatform.customplugin.springboot;
 
 import eu.miaplatform.service.environment.EnvConfiguration;
 import eu.miaplatform.service.environment.InvalidEnvConfigurationException;
-import io.swagger.annotations.ApiOperation;
 import javafx.application.Application;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,17 +10,17 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import springfox.documentation.annotations.ApiIgnore;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.PostConstruct;
 import javax.servlet.http.HttpServletRequest;
 
+import static eu.miaplatform.customplugin.springboot.CPConstants.CP_REQUEST;
+
 @RestController
 public abstract class CPController {
 
-    protected final String CP_REQUEST = "CP_REQUEST";
     protected final Logger logger = LoggerFactory.getLogger(CPController.class);
 
     @Autowired
@@ -42,26 +41,5 @@ public abstract class CPController {
     @ModelAttribute(CP_REQUEST)
     public CPRequest populate(HttpServletRequest request) {
         return new CPRequest(request, new Options());
-    }
-
-    @GetMapping("/-/healthz")
-    @ApiOperation(value = "Healthz")
-    @ResponseBody
-    public ResponseEntity healthz(@ApiIgnore @ModelAttribute(CP_REQUEST) CPRequest cpRequest) {
-        return healthinessHandler(cpRequest);
-    }
-
-    @GetMapping("/-/ready")
-    @ApiOperation(value = "Ready")
-    public ResponseEntity ready(@ApiIgnore @ModelAttribute(CP_REQUEST) CPRequest cpRequest) {
-        return readinessHandler(cpRequest);
-    }
-
-    public ResponseEntity healthinessHandler(CPRequest cpRequest) {
-        return CPStatus.statusOk(new CPStatusBody());
-    }
-
-    public ResponseEntity readinessHandler(CPRequest cpRequest) {
-        return CPStatus.statusOk(new CPStatusBody());
     }
 }
